refactor: move settings in mail layout refactor: create sidebar directory and remove account-switcher fix: mail-compose-modal styles
