+++ app/(routes)/mail/layout.tsx
@@ -1,4 +1,4 @@
-import { AppSidebar } from "@/components/ui/app-sidebar";
+import { AppSidebar } from "@/components/mail/sidebar/app-sidebar";
 
 export default function MailLayout({ children }: { children: React.ReactNode }) {
   return (
