+++ app/(routes)/mail/settings/shortcuts/page.tsx
