+++ components/mail/mail-compose-modal.tsx
@@ -18,20 +18,15 @@ export default function MailComposeModal() {
 
   return (
     <Credenza open={isOpen} onOpenChange={setIsOpen}>
-      <CredenzaContent>
+      <CredenzaContent className="p-0">
         <VisuallyHidden>
           <CredenzaHeader>
             <CredenzaTitle>Compose</CredenzaTitle>
           </CredenzaHeader>
         </VisuallyHidden>
-        <CredenzaBody>
+        <CredenzaBody className="hide-scrollbar overflow-y-auto px-0 py-0">
           <MailCompose onClose={close} />
         </CredenzaBody>
-        <CredenzaFooter>
-          <CredenzaClose asChild>
-            <button>Close</button>
-          </CredenzaClose>
-        </CredenzaFooter>
       </CredenzaContent>
     </Credenza>
   );
